coord for mrconvert
mrconvert crashes if -coord is float

+++ nipype/interfaces/mrtrix3/utils.py
@@ -560,7 +560,7 @@ class MRConvertInputSpec(MRTrix3BaseInputSpec):
         desc="output image",
     )
     coord = traits.List(
-        traits.Float,
+        traits.Int,
         sep=" ",
         argstr="-coord %s",
         desc="extract data at the specified coordinates",
